tests/wpt/web-platform-tests/css/css-transforms/parsing/backface-visibility-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transform Module Level 2: getComputedStyle().backfaceVisibility</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#backface-visibility-property">
<meta name="assert" content="backface-visibility computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("backface-visibility", "visible");
test_computed_value("backface-visibility", "hidden");
</script>
</body>
</html>
